feat: add new gitlab approval api handling
Implemented a new feature that allows users to use the GitLab Merge
Request Approval feature (only available in GitLab Starter and Bronze).

For more information about the Merge Request Approval feature, go to
https://docs.gitlab.com/ee/user/project/merge_requests/merge_request_approvals.html

Add new configuration to choose between "upvotes" or "approvals". This
is a global setting, that means that it will be the same for all the
instances. I think I'll implement a way to handle this setting per
repository in the future. I'm open to suggestion and Pull Request.

Replace two existing configurations so they now respect the VS Code
extension development conventions.

At the moment, the old one are deprecated, so they are only visible in
the JSON, while the new ones are actually available in the UI.

Feel free to update to the new ones as the old ones will be removed in
the next release.

Enjoy 🎉